--- a/env.common.yml
+++ b/env.common.yml
@@ -18,6 +18,7 @@ dependencies:
   - nbsphinx
   - pandoc
   - black
+  - numba
   - pip:
     - demjson
     - Pillow