This PR implements a simple load balancing algorithm inside the batch sampler. The new `BalancedBatchSampler` either balances by number of atoms or by number of neighbors. Load balancing is activated by default (in atoms mode). Balancing relies on metadata stored alongside the lmdbs. If the `metadata.npz` file is not found, we print a warning. It is then deactivated per default. However, if the user explicitly specifies a load balancing mode it stays activated. We then load the data in the sampler to access its size, which incurs some overhead.

## Motivation
The size of a data sample can vary by up to an order of magnitude, which currently leads to skewed usage of both GPU memory and compute. In the worst case a GPU can draw an unlucky batch and run out of memory even 100k steps after training started.

For 16x8 GPUs with per-GPU batch size 4 (batch size 512 in total), this PR leads to a speed-up of around 3x and a memory reduction of more than 50%. For 8x8 GPUs with per-GPU batch size 32 (batch size 2048 in total), it leads to a speed-up of more than 20%.

## Considerations
The batch sampler runs in the main thread, not in a worker process. So this code is performance critical. Possible performance improvements:
- [x] numba for `balanced_partition` function: 100ms -> 0.5ms, no caching to prevent file system errors
- [x] add lmdbs only containing sizes (natoms and neighbors), so we don't have to load the data twice: 80ms -> 0.02ms
- use a different communication method, so the batch samplers don't push data to the GPU: 15ms, includes synchronization barrier (80ms before optimized loading). Probably not worth it.

Times measured using 16x8 GPUs with per-GPU batch size 4 (batch size 512 in total).

## Other changes
- We now set `eval_batch_size` to `batch_size` per default, instead of an arbitrary 64.
- Remove `args.num_workers`, since this is a duplicate of `config["optim"]["num_workers"]`
- add `numba` as a dependency
- Fix inconsistent validation intervals by using `self.step` instead of `iters`