@abhshkdz 2021-09-14
Let's open a separate PR to add a script to generate metadata files for S2EF, and package them in the LMDBs for IS2*